Assignment3_OMPL_Docker_benchmark_planners/requirements.txt
pyyaml >= 6.0.0
numpy < 2
meshcat >= 0.3.2
python-fcl >= 0.7
https://github.com/ompl/ompl/releases/download/prerelease/ompl-1.6.0-cp312-cp312-manylinux_2_28_x86_64.whl